Add initial context file.
This lays the foundation for eventually turning this metadata into VCs.

The context uses a mix of the W3Cs EARL vocabulary, DOAP (which EARL
uses), Schema.org, and a temporary `reporter` namespace mapped to
https://github.com/digitalbazaar/mocha-w3c-interop-reporter#
which is the primary software which consumes these details to generate
W3C test reports.

This context MAY (and likely should) move in the future either to that repo or
to some more permanent home.